Given an image, predict what clothing category it falls under amongst the 10 listed categories. Fashion MNIST dataset used.
What the code does is- Given an image, predict what clothing category it falls under amongst the 10 listed categories. The datset used is the Fashion MNIST .
About the dataset: One of the foundational datasets for learning and benchmarking algorithms is the Modified National Institute of Standards and Technology (MNIST) database, by Yann LeCun, Corinna Cortes, and Christopher Burges. This dataset is comprised of images of 70,000 handwritten digits from 0 to 9. The images are 28 × 28 grayscale.